Making GIBBON robust to numerical instability (#1814)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1814

When using GIBBON for batch acquisition, numerical instabilities can be encountered that lead a quantity (`V_determinant`) that is mathematically positive to be numerically negative. Taking the logarithm of this quantity then leads to NaNs in the information gain tensor, which in turn leads to optimization failures.

This commit clamps a term in the expression for `V_determinant` so that the resultant quantity is positive, while ensuring that gradients continue to flow through the backward pass to all relevant variables.

Reviewed By: Balandat

Differential Revision: D45584690

fbshipit-source-id: de464213d8fa10f8bc7000e584d22ccc5ffbe641