Make SCBO tutorial deterministic (#1833)
Summary:
## Motivation

The SCBO tutorial is not deterministic and uses a while loop, so its runtime can vary quite a bit. I saw 50s and 100s in smoke test mode on my laptop. In the CI, it usually runs well under the 5-minute limit in smoke test mode, but recently timed out and caused a failure.

This change makes it deterministic by seeding the SobolEngine.

Pull Request resolved: https://github.com/pytorch/botorch/pull/1833

Test Plan: Ran the tutorial to make sure it runs and is deterministic.

Reviewed By: Balandat

Differential Revision: D45830428

Pulled By: esantorella

fbshipit-source-id: adb4c8a351f33f8501252439e62cbb8c392a7bdd