Fix flaky VariationalGP unit test (#1836)
Summary:
## Motivation

A newly-added test in `TestSingleTaskVariationalGP` is flaky; a numerical check can fail depending on the seed. The failure is not surprising because the test is asserting that the mean posterior mean over the train data equals the mean of the `y` data.

### Have you read the [Contributing Guidelines on pull requests](https://github.com/pytorch/botorch/blob/main/CONTRIBUTING.md#pull-requests)?

Yes

Pull Request resolved: https://github.com/pytorch/botorch/pull/1836

Test Plan: Ran it with 100 different seeds

Reviewed By: Balandat

Differential Revision: D45885496

Pulled By: esantorella

fbshipit-source-id: c5f48aa67f72ebd86219f741bf84a29e2faf5b07