Refactor test function test utils (#1839)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1839

Refactors the test utils for botorch test functions to make things more modular.

Instead of subclassing the base test cases, we now have class Mixins for synthetic, constrained, and multi-objective problems. This means we can mix-and-match the test case as needed.

This also flattens the `TestMultiObjectiveProblems` and `TestConstrainedMultiObjectiveProblems` test collections (the idea behind the test design is that we have separate test cases for each problem (potentially with different arguments) to better identify and group the tests, so this reestablishes that.

Reviewed By: esantorella

Differential Revision: D45969036

fbshipit-source-id: dbb694536aff9c844fc7060f115aedfbb16a2575