Allow passing in task features as part of X in MTGP.posterior (#1868)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1868

Prior to this change, `MultiTaskGP` could be evaluated using inputs of shape `batch x d` when using the model directly, and using inputs of shape `batch x d + 1` when the model was wrapped in a `ModelListGP`. This diff (combined with https://github.com/pytorch/botorch/pull/1854) legitimizes both input shapes regardless of what API is used to evaluate the model. This will unify the APIs and increase the flexibility in how the model is used.

Differential Revision: https://internalfb.com/D46496775

fbshipit-source-id: e79dee0cc0ba0cd03fb97382cc1eb9ef4e5a8371