Change how ModelListGP.posterior evaluates sub-models (#1854)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1854

Prior to this change, `ModelListGP.posterior` would call it's submodels directly and work with the returned MVNs, skipping any processing in the individual `posterior` methods. This was particularly consequential for the `MultiTaskGP` (and subclasses) where skipping the `posterior` meant changing the expected shape of the `X` required by the `posterior`.

With this change, `ModelListGP.posterior` evaluates the `posterior` methods for each of its submodels, and combines the MVNs into a single MVN where applicable. For `MultiTaskGP`, this makes it so that the `posterior` can be evaluated using the same inputs regardless of whether the model is wrapped in a `ModelListGP` or not.

Reviewed By: Balandat

Differential Revision: D46364187

fbshipit-source-id: ee3d5fa738c80f1bb77771789cb1d287c9034251