Stop round transform test from being flaky (#1903)
Summary:
## Motivation

A test for the round transform asserting that approximately-rounded inputs are closer to exactly-rounded values than the original inputs are fails flakily. This happens because an error is introduced due to the approximation used:

```python
x = torch.tensor(0., dtype=torch.float64)
# tensor(4.5398e-05, dtype=torch.float64)
# equivalent to (tanh(-5) + 1) / 2
print(approximate_round(x, tau=1e-1))
```

This PR sets the tolerance for the relevant test to the error induced by the approximation. I also added to a docstring.

### Have you read the [Contributing Guidelines on pull requests](https://github.com/pytorch/botorch/blob/main/CONTRIBUTING.md#pull-requests)?

Yes

Pull Request resolved: https://github.com/pytorch/botorch/pull/1903

Test Plan: Is a test.

Reviewed By: saitcakmak

Differential Revision: D47027914

Pulled By: esantorella

fbshipit-source-id: ac0e0006098f0622ea8750ec81d9e94feecd8c6f