speed up LCE-A kernel (#1910)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1910

X-link: https://github.com/facebook/Ax/pull/1694

The current implementation is very slow. This is particularly problematic when the number of contexts is large.

This diff yields huge speed ups (>3 orders of magnitude) by removing nested for loops in `forward` and instead using batched computation. This also caches the context_covar in eval mode, with an additional requirements that the parameters for each context are contiguous and in the same order.

* 14x speed up with 8 contexts and 1,154x speed up with 128 contexts (CPU)

* 22x speed up with 8 contexts and 3,370x speed up with 128 contexts (CUDA)
* Without this diff, it takes 5 seconds on CPU and 12 seconds with CUDA for a single forward pass with 128 contexts.

Current implementation:

* 8 contexts:
  * Forward pass: 20ms (CPU), 45.2ms (CUDA)
  * Roundtrip: 39.1ms (CPU), 99.7ms (CUDA)
* 128 contexts:
  * Forward pass: 5.08s (CPU), 12s (CUDA)
  * Roundtrip: 14.2s (CPU), 26.7s (CUDA)

New Implementation:
* 8 contexts:
  * Forward pass: 1.44ms (CPU), 2.05ms (CUDA)
  * Roundtrip: 2.22ms (CPU), 4.65ms (CUDA)
* 128 contexts:
  * Forward pass: 4.4ms (CPU), 3.56ms (CUDA)
  * Roundtrip: 6.97ms (CPU), 5.34ms (CUDA)

Reviewed By: Balandat

Differential Revision: D47118335

fbshipit-source-id: 27f87ec908f64d7380b2596d7c969c9d1801b9d5