Batch-MC Prior-Guided AF (#1925)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1925

see title. This weights the AF values before reduction over the sample and q-dims.

This also handles `X_pending` directly, rather than on the underlying AF. This ensures, `X_pending` is concatenated with `X` and passed to both the AF an the prior. This will only work for AFs that support concatenating `X_pending`.

This would need revision to work with for example NEHVI with cached box decompositions.

Reviewed By: Balandat

Differential Revision: D47307059

fbshipit-source-id: 0f7567a6036928c4df82068d809c947fb0f92a61