Decoupled Acquisition Function (#1948)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1948

Introduce an abstract class for decoupled acquisition functions.

A decoupled acquisition function where one may intend to evaluate a design on only a subset of the outcomes. Typically this would be handled by fantasizing, where one would fantasize as to what the partial observation would be if one were to evaluate a design on the subset of outcomes (e.g. you only fantasize at those outcomes)

Differential Revision: https://internalfb.com/D47710904

fbshipit-source-id: 1dbebeaa4fac09a74f25cfc441bef4dcf5fb7d3f