add support for decoupled evaluations in cost-aware utilities (#1949)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1949

This adds basic support for decoupled evaluations in cost-aware utilities. This is "basic" in the sense that it only supports scenarios where all `q` points in `X` are evaluated on the same output(s). This is sufficient for scenarios where either `q=1`, all `q` points are evaluated on the same output(s), or the asynchronous setting where the cost is only evaluated on the new point (where effectively `q=1`) and not the pending points.

Differential Revision: D47710947

fbshipit-source-id: 706d7054db3fbd506bbd3891a6892e1093638ef5