Merge SupervisedDataset & FixedNoiseDataset (#1945)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1945

This diff deprecates `FixedNoiseDataset` and merges it into `SupervisedDataset` with Yvar becoming an optional field. This also simplifies the class hierarchy a bit, removing `SupervisedDatasetMeta` in favor of an `__init__` method.

I plan to follow up on this by adding optional metric names to datasets and introducing a MultiTaskDataset, which will simplify some of the planned work in Ax MBM.

Reviewed By: esantorella

Differential Revision: D47729430

fbshipit-source-id: 551cd78a02755505573b10ea1f075aa21f838ab7