Fix fantasization with FixedNoiseGP and outcome transforms and use FantasizeMixin (#2011)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2011

This fixes fantasization with FixedNoiseGP when using outcome transforms----previously, already-transformed noise was transformed again during fantasization.

This also improves the fantasization for batched and batched multi-output models to use the average noise for each batch and output.

This also removes repeated code and uses the logic in `FantasizeMixin.fantasize` for handling `X` with size 0 on the -2 dimension.

This also deprecates the use of `observation_noise` as a boolean argument to fantasize.

Differential Revision: https://internalfb.com/D49200325

fbshipit-source-id: 686663284452f02114695d1bb2da973a16c3267e