modify `PenalizedMCObjective` to support non-batch eval (#2073)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2073

To match penalty term with MC objective, we current unsqueeze the first dim which corresponds to the dimension of MC samples. However, when a `qxd`-dim X tensor is evaluated e.g. computing feasibility, it causes shape mismatch. As one would expect `q`-dim tensor returned, it will return `1xq`-dim tensor instead.

To fix, we check the dims of obj; if it is non-mc samples, we will sequeeze the first dim back.

Reviewed By: bletham

Differential Revision: D49305807

fbshipit-source-id: c8e2dedbdb38d79b5910aca393cf2d3d4e1e311f