Adding `posterior_transform` to `qMaxValueEntropySearch`'s input constructor (#2181)
Summary:

This diff adds `posterior_transform` to the input constructor of the `qMaxValueEntropySearch` acquisition function. This can be used to transform the posterior distribution before computing the maximum entropy search, a property that Ax relies on to encode whether an objective is to be maximized or minimized, instead of passing the `minimize` keyword, which is not accepted by  all acquisition functions.

This solves https://github.com/facebook/Ax/issues/2133.

Differential Revision: D53101380