Do not normalize dimension if all values are equal
Summary:
Suppose that the train data has x0 as a constant (can happen with few data points) but it is being optimized in [0, 1]. In the current behavior, we first calculate a coefficient of 0.0, then clamp this up to 1e-8. During acqf optimization, we will evaluate the model with values in [0, 1], which will then get normalized to [0, 1e8]. This can cause numerical issues in GPyTorch and lead to non-PSD covariance matrices since the model was trained with constant inputs, likely learning much more reasonable lengthscales that don't play well with these large values.

This diff adds a check if the learned range is less than `almost_zero`, it treats the corresponding dimension as constant and uses a coefficient of 1.0 instead.
RFC: An alternative would be to simply change the behavior of `min_range` and use it for this purpose rather than rounding up.

Reviewed By: esantorella

Differential Revision: D53213759