unify ModelList and ModelListGP subset_output behavior
Summary:
`ModelList` and `ModelListGP` have different `subset_output` methods. This is bug prone and is bug in SEBO.

`ModelList.subset_output` considers that `Model`s in the list may themselves be multi-output, and considers that when splitting the input argument `idcs` across models in the list---idcs is used to index into outputs

`ModelListGP.subset_output` does not consider that sub models may be multi-output and instead applies `idcs` is used to index into models.

A significant issue here is that `MultitaskGP` has the number of outputs equal to the number of tasks. This causes issues for typical uses cases where all tasks refer to the same Ax outcome e.g. the real metric and a proxy of it. This is not an issue with the current use of `ModelListGP`, but it is with `Modellist`.

To fix this, this diff unifies the `subset_output` behavior for `ModelListGP` and `Modellist` and prepares the appropriate arguments in Ax.

Ax-side diff will follow.

Reviewed By: saitcakmak, Balandat

Differential Revision: D54203833

fbshipit-source-id: 2228d1cf1b7b758d6c7fbaa7b36ae5e9e0c01849