Fix non-None constraint noise levels in some constrained test problems (#2241)
Summary:
This was inadvertently overridden / nnot properly set in some constructors.

We currently don't properly test that this attribute is correctly assigned upon instantiation. It's not straightforward to figure out how to easily do this in the current test setup; we can figure that out in a follow-up.

Pull Request resolved: https://github.com/pytorch/botorch/pull/2241

Reviewed By: SebastianAment

Differential Revision: D54669269

Pulled By: Balandat

fbshipit-source-id: 18248c3851b1235a98f08389215e8e5cc5ffbe74