More removal of unused ** arguments (#2336)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2336

I am auditing cases in which BoTorch functions admit ** arguments and then don't use them. In many cases, it was easier to fix the issue than write it up. These are the easy cases.

Note on inheritance: Type-checkers say that if a method accepts an argument, so must methods that override it.
* I removed `**kwargs` from abstract methods to remove the implication that their subclasses must also support `**kwargs` even if they don't use them.
* In cases where the base method admits `**kwargs` and is GPyTorch, I chose to "inconsistent override" GPyTorch rather than also having ignored `**kwargs` in BoTorch. This was the case for overriding `Module.forward` (in an ExactGP), `Kernel.forward`, and `Likelihood.forward`.

Changes:
* Small correctness fix in error-catching
* Some small typing fixes
* Removed `**kwargs` in many cases.

Reviewed By: saitcakmak

Differential Revision: D56849296

fbshipit-source-id: b059148e018608fac9691ee255cd73118d2e52b1