Explicitly check timeout against None so that 0.0 isn't ignored
Summary: While testing some timeout-related code, I set a timeout to 0.0 to ensure that the timeout would bind. I was confused when nothing happened; it turned out that the presence of a timeout was tested for with "if timeout" rather than "if timeout is None," so that 0.0 was not considered a timeout.

Reviewed By: Balandat, SebastianAment

Differential Revision: D57542718