Explicitly check timeout against None so that 0.0 isn't ignored (#2348)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2348

While testing some timeout-related code, I set a timeout to 0.0 to ensure that the timeout would bind. I was confused when nothing happened; it turned out that the presence of a timeout was tested for with "if timeout" rather than "if timeout is None," so that 0.0 was not considered a timeout.

Reviewed By: Balandat, SebastianAment

Differential Revision: D57542718

fbshipit-source-id: 9de3bbb20e9c21ada28d6357c7ef64dcbf829f3f