support evaluating posterior predictive in MultiTaskGP (#2375)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2375

Previously, evaluating the posterior predictive was not supported in MultitaskGP. This adds support for using the average noise per task. Note currently MultitaskGP with an inferred noise level only infers a single noise level that is shared across tasks, so this is change is most useful when the noise is observed.

Reviewed By: saitcakmak

Differential Revision: D58227605

fbshipit-source-id: 515b2240453ef16694257514442d71f1162a9221