Deprecate model conversion code (#2431)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2431

These utilities are scarcely used and poorly maintained. They also are not fully compatible with different GPyTorch priors that we plan to use by default in the near future. Marking them for deprecation in v0.13.

Also added an explicit warning when used with priors that do not have any state dict, since such priors cannot be verified to be compatible across submodels.

Reviewed By: SebastianAment

Differential Revision: D59813960

fbshipit-source-id: 8e2fa45debc3f801e45d88e62b4d31d5661c22e1