Simplify qMultiObjectiveMaxValueEntropy.__init__
Summary:
I tried to see if I could eliminate the use of model converter code in `qMultiObjectiveMaxValueEntropy` with a reasonable time investment, but couldn't go much further than this.

The in-line comments about tensor shapes do not appear to be fully accurate, particularly those in `qMaxValueEntropy._compute_information_gain`. They seem to have been written for the multi-fidelity case and are confusing when viewed without that context. I may revisit this later.

Differential Revision: D59834357