Simplify qMultiObjectiveMaxValueEntropy.__init__ (#2433)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2433

I tried to see if I could eliminate the use of model converter code in `qMultiObjectiveMaxValueEntropy` with a reasonable time investment, but couldn't go much further than this.

The in-line comments about tensor shapes do not appear to be fully accurate, particularly those in `qMaxValueEntropy._compute_information_gain`. They seem to have been written for the multi-fidelity case and are confusing when viewed without that context. I may revisit this later.

Reviewed By: SebastianAment

Differential Revision: D59834357

fbshipit-source-id: 8b70f4dc048825517341cbc1f6897f1f52cc6936