Add a helper for producing a DeterministicModel using a Matheron path (#2435)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2435

Out of the box, pathwise sampling code does not produce outputs that conform to BoTorch modeling conventions. This adds a helper that produces a Matheron path and wraps it in a `GenericDeterministicModel` using a helper that reshapes the path outputs to mimic the expected behavior from the model's posterior.

This will offer a straightforward replacement of `get_gp_samples`, which is being deprecated.

Reviewed By: esantorella

Differential Revision: D59941730