Unpin numpy from <2.0 (#2459)
Summary:
This unpins numpy from <2.0, which was done as a temporary workaround in https://github.com/pytorch/botorch/issues/2382 to avoid issues with windows builds.

The underlying issue was an incompatibility between pytorch 2.4.0 windows builds and numpy 2.0. Pointed pip to the updated pytorch pip index to pick up newer 2.5.0 dev versions to resolve the issue: https://github.com/pytorch/botorch/pull/2459#issuecomment-2267577174

There is still the issue of existing torch release windows binaries potentially not being compatible with numpy 2.0, but that shouldn't be something that we need to handle.

Pull Request resolved: https://github.com/pytorch/botorch/pull/2459

Reviewed By: esantorella

Differential Revision: D60735276

Pulled By: Balandat

fbshipit-source-id: 5af618222a6ad8acbedfecc3a7ca4f8a6b45d098