Suppress non-log EI warnings in HVKG helpers (#2486)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2486

My first idea was to change this to use Log-EHVI, but this helper is used in the tutorial to compute the hypervolume, so we can't change the return value. Since this is not directly constructed by the user, raising a warning is not productive. This diff suppresses the warning.

Differential Revision: D61722790