Use mathjax svg renderer (#2481)
Summary:
We identified an issue where mathjax would fail to render certain math symbols. This happens on MacOS versions Ventura and later (13+) because of a change in the OS-supplied STIX fonts. We fix this here by changing the renderer from HTML+CSS (which tries to use local STIX fonts) to SVG.

More details on the issue can be found here: https://meta.stackexchange.com/a/389308

| **Before**    | <img width="933" alt="image" src="https://github.com/user-attachments/assets/85a381bf-fb43-46b6-a90b-84a2567b7e30"> |
| -------- | ------- |
| **After**  | <img width="933" alt="image" src="https://github.com/user-attachments/assets/0121a602-a8ad-4f44-b68b-fded891bf04a">    |

### Considerations

The [official docs](https://docs.mathjax.org/en/stable/output.html#mathjax-output-formats) outline a couple considerations for using the SVG renderer, the biggest being lack of IE8 support. If IE8 support is crucial then we should consider alternate approaches.

An alternative fix could be to upgrade our mathjax version as later versions may not be as reliant on local STIX fonts but the 3.0 update was a complete rewrite and appears to [not be a straightforward upgrade](https://docs.mathjax.org/en/latest/upgrading/v2.html).

Pull Request resolved: https://github.com/pytorch/botorch/pull/2481

Reviewed By: Balandat

Differential Revision: D61681247

Pulled By: CristianLara

fbshipit-source-id: 8842997e140722d8e3577b734e6d42cf71caac03