Make BaseTestFunction.evaluate_true accept 1d inputs (#2492)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2492

Context:

Currently, every test function except for `AugmentedBranin` has an `evaluate_true` method that works with 1d inputs. It is actually surprising that so many work, since `BaseTestFunction` is currently written so that `BaseTestFunction.forward` casts inputs to 2d before passing them to `BaseTestFunction.evaluate_true`. So currently, it's not clear if we should expect `evaluate_true` to work with 1d inputs, but nonetheless this is happening downstream.

This PR:
* Requires `evaluate_true` to work with 1d inputs
* Removes the logic that expands the dimension of unbatched tensors in `forward` before passign to `evaluate_true` and then removes the batch dimension in favor of leaving unbatched tensors unbatched everywhere
* Changes `AugmentedBranin` to work with 1d inputs
* Fixes a couple type errors
* Expands docstrings

Reviewed By: saitcakmak

Differential Revision: D61916387

fbshipit-source-id: 54690847fbba5d3155294560219a67cbf498919d