Fix input constructor for `qMultiFidelityKnowledgeGradient` (#2519)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2519

The input constructor for `qMultiFidelityKnowledgeGradient` was computing `current_value` in a manner that was effectively not fidelity-aware (the fidelity dimension was not fixed in the optimization done to obtain `current_value`), since it was simply re-using functionality from the input constructor for `construct_inputs_qKG`.

This diff addresses this by making use of `target_fidelities` as the `fixed_features` argument to `D62380369`.

Differential Revision: D62391106