Fix input constructor for `qMultiFidelityKnowledgeGradient` by fixing fidelity dimensions when computing `current_value` (#2519)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2519

The input constructor for `qMultiFidelityKnowledgeGradient` was computing `current_value` in a manner that was effectively not fidelity-aware (the fidelity dimension was not fixed in the optimization done to obtain `current_value`), since it was simply re-using functionality from the input constructor for `construct_inputs_qKG`.

This diff addresses this by making use of `target_fidelities` as the `fixed_features` argument to `D62380369`.

Reviewed By: Balandat

Differential Revision: D62391106

fbshipit-source-id: acf4f4a9a3644bb467decac70ce38a39c1c9ffc8