Apply input transforms when computing MLL in model closures (#2527)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2527

During model training, the input transforms are applied in `model.forward`. While evaluating the model closures, we pass in the train inputs to the `mll`, which passed them down to the `likelihood`. If we don't transform the inputs before passing them into `mll`, we end up evaluating `model.forward` and `likelihood` using different inputs.

This is not an issue during the `posterior` evaluation, since the transforms are applied in `model.posterior` before being passed to `model.__call__` and `likelihood`.

This diff updates the model closures to transform the inputs before passing them into `mll`.

Fixes https://github.com/pytorch/botorch/issues/2515

Reviewed By: SebastianAment

Differential Revision: D62497392

fbshipit-source-id: 9850c6529eea336589c2e2bbae400a4a9dc87f12