OptimizationGradientError (#2537)
Summary:
<!--
Thank you for sending the PR! We appreciate you spending the time to make BoTorch better.

Help us understand your motivation by explaining why you decided to make this change.

You can learn more about contributing to BoTorch here: https://github.com/pytorch/botorch/blob/main/CONTRIBUTING.md
-->

## Motivation

I am currently experimenting with some custom kernels, which exhibit some numeric instabilities, which lead in some situations during optimization to `NaN`s in the gradient. To better examine these situations, I thought it could be nice if the error which is thrown in this situation also contains the current x which leads to the `NaN`s in the gradient. For this purpose, I implemented an `OptimizationGradientError` like the `OptimizationTimeoutError` which holds the current x.

Im also thinking about catching the error in `timeout.py` as the `OptimizationTimeoutError`, throwing a warning and returning a result of the optimization. What do you think about this? I think this could be a useful and elegant way of dealing with it.

### Have you read the [Contributing Guidelines on pull requests](https://github.com/pytorch/botorch/blob/main/CONTRIBUTING.md#pull-requests)?

Yes.

Pull Request resolved: https://github.com/pytorch/botorch/pull/2537

Test Plan: Unit tests.

Reviewed By: esantorella

Differential Revision: D62710945

Pulled By: Balandat

fbshipit-source-id: c69cae3bc2cd3d11a763843c3db208d3b63a7124