Apply PEP 604 union type syntax codemod (#2561)
Summary:
This codemods all `Optional[X]` type definitions to use the PEP 604 syntax `X | None` instead. I also ran ufmt to fix the imports.

Why? I came across https://github.com/asottile/pyupgrade randomly and figured I'd give it a shot...

This is mostly to test that everything runs smoothly. This may be kind of a pain to rebase on so not sure we should merge this in in one big chunk...

Pull Request resolved: https://github.com/pytorch/botorch/pull/2561

Reviewed By: esantorella

Differential Revision: D63733395

Pulled By: Balandat

fbshipit-source-id: 58e700487cb63223a4bc54a692567f0debd8f8a7