Introduce `fast_optimize` context manager to speed up testing (#2563)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2563

Context: Many BoTorch tests take a while because they run optimization code. But it's good that this code runs rather than being avoided or mocked out, becuase the tests are ensuring that things work end-to-end. Borrowing a page from Ax's `fast_botorch_optimize`, this commit introduces the same thing to BoTorch, with the exception of `fit_fully_bayesian_model_nuts`.

A future commit to Ax can remove that functionality from Ax in favor of importing it from BoTorch, but we might not want to do it right way because then Ax won't work with older versions of BoTorch.

This PR:
* Introduces `fast_optimize`, which is the same as Ax's `fast_botorch_optimize`, but with different import paths.
* Applies it to a slow test, reducing runtime to 2s from 6s-10s.

Reviewed By: sdaulton

Differential Revision: D63838626

fbshipit-source-id: d2f8d6b496df1b50baf4f07327713a38473157d4