Make input & outcome transforms first class attributes of Model
Summary: These were previously only set when they were not None, which lead to a lot of `hasattr`, `getattr` usage throughout the codebase to check for them. This diff adds them as attributes to base `Model` class with default values of `None`. With this change, we can now access `model.input/outcome_transform` on all models.

Differential Revision: D66012223